+++ src/NORS.Plugin/NorsPlugin.cs
@@ -10,7 +10,7 @@ namespace NORS.Plugin
     public class NorsPlugin : BaseUnityPlugin
     {
         public const string Guid = "com.dsr.nors";
-        public const string Version = "0.7.7";
+        public const string Version = "0.7.8";
 
         internal static NorsPlugin Instance;
         internal static ManualLogSource Log;